Add support for unencrypted description field
When the description field is not encrypted, the resource secrets will contain only a password instead of a JSON containing password and description. Fixes #9 by detecting if the decrypted resource secrets is a JSON or not and act accordingly.
The description field will be extracted from the resource secrets if present, otherwise it will be retrieved from the unencrypted resource details.